project2/simple approach/requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.3.9
gast==0.2.2
grpcio==1.21.1
h5py==2.7.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
nltk==3.2.5
numpy==1.16.4
pandas==0.22.0
protobuf==3.8.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.19.1
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Theano==1.0.4
Werkzeug==0.15.4
